Fix pull to only run when changes are detected

+++ ellipsis.sh
@@ -10,14 +10,18 @@ pkg.install() {
 }
 
 pkg.pull() {
-    # Unlink old files
-    hooks.unlink
+    # Check for updates on git
+    git remote update 2>&1 > /dev/null
+    if git.is_behind; then
+        # Unlink old files
+        hooks.unlink
 
-    # Pull package changes
-    git.pull
+        # Pull package changes
+        git.pull
 
-    # Link new files
-    pkg.link
+        # Link new files
+        pkg.link
+    fi
 
     # Install docker
     bash $PKG_PATH/run.sh "$ELLIPSIS_SRC"
